feat: add acr validation check (#23)
